@@ -10,16 +10,17 @@
         }
     ],
     "require": {
-        "php": ">=5.4.0",
-        "voryx/thruway": "0.3.*",
-        "illuminate/support": "5.*",
-        "illuminate/console": "5.*"
+        "php": ">=7.0.1",
+        "illuminate/support": "^5.5",
+        "illuminate/console": "^5.5",
+        "thruway/client": "^0.5.3",
+        "thruway/pawl-transport": "^0.5.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.*",
+        "phpunit/phpunit": "^6",
         "mockery/mockery": "0.9.*",
-        "fabpot/php-cs-fixer": "*",
-        "squizlabs/php_codesniffer": "1.*"
+        "squizlabs/php_codesniffer": "^3.2",
+        "friendsofphp/php-cs-fixer": "^2.11"
     },
     "autoload": {
         "psr-0": {
